Add API endpoint to reset database
And move the fixtures module to the root directory.
