+++ backend/src/bin/api.rs
@@ -3,13 +3,14 @@ use std::net::{Ipv4Addr, SocketAddr};
 use std::path::PathBuf;
 use std::str::FromStr;
 
+#[cfg(feature = "dev-database")]
+use backend::fixtures;
+use backend::router;
 use clap::Parser;
 use sqlx::sqlite::SqliteConnectOptions;
 use sqlx::SqlitePool;
 use tokio::net::TcpListener;
 use tower_http::services::{ServeDir, ServeFile};
-
-use backend::router;
 
 /// Abacus API server
 #[derive(Parser, Debug)]
@@ -83,46 +84,4 @@ async fn create_sqlite_pool(
     }
 
     Ok(pool)
-}
-
-#[cfg(feature = "dev-database")]
-mod fixtures {
-    /// Macro to convert a single fixture name to the contents of a fixture file
-    macro_rules! load_fixture {
-        ($fixture:literal) => {
-            Fixture {
-                data: include_str!(concat!("../../fixtures/", $fixture, ".sql")),
-            }
-        };
-    }
-
-    /// Macro to convert the list of fixtures to their contents
-    macro_rules! load_fixtures {
-        ([$($fix:literal),* $(,)?]) => {
-            &[$(load_fixture!($fix),)*]
-        }
-    }
-
-    /// List of fixtures to load when data seeding is requested.
-    ///
-    /// This list should be updated manually when a new fixture is added that
-    /// needs to be loaded when seeding fixtures.
-    const FIXTURES: &[Fixture] = load_fixtures!(["elections", "polling_stations",]);
-
-    /// The data contained in a fixture file
-    struct Fixture {
-        data: &'static str,
-    }
-
-    /// Function that loads the fixture data into the given connection
-    /// Each fixture may contain multiple SQL statements.
-    pub async fn seed_fixture_data(
-        pool: &sqlx::SqlitePool,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        for fixture in FIXTURES {
-            sqlx::raw_sql(fixture.data).execute(pool).await?;
-        }
-
-        Ok(())
-    }
 }
